+++ nf_core/modules/modules_json.py
@@ -253,7 +253,7 @@ class ModulesJson:
                         break
                     # Create a new modules repo with the selected branch, and retry find the sha
                     modules_repo = nf_core.modules.modules_repo.ModulesRepo(
-                        remote_url=remote_url, branch=branch, no_pull=True, no_progress=True
+                        remote_url=remote_url, branch=branch, no_pull=True, hide_progress=True
                     )
                 else:
                     found_sha = True