+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Replace use of the deprecated `distutils` Version object with that from `packaging` ([#1735](https://github.com/nf-core/tools/pull/1735))
 - Add code to cancel CI run if a new run starts ([#1760](https://github.com/nf-core/tools/pull/1760))
 - CI for the API docs generation now uses the ubuntu-latest base image ([#1762](https://github.com/nf-core/tools/pull/1762))
+- Add option to hide progress bars in `nf-core lint` and `nf-core modules lint` with `--hide-progress`.
 
 ### Modules